% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_outputs_gh_url}
\alias{get_outputs_gh_url}
\title{Get the direct URL to a file in the NHP Outputs GitHub repo}
\usage{
get_outputs_gh_url(...)
}
\arguments{
\item{...}{Pass the name of the file in via \code{...}}
}
\description{
Get the direct URL to a file in the NHP Outputs GitHub repo
}
\keyword{internal}
