% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_su_gh_file}
\alias{get_su_gh_file}
\title{Read in a file from a Strategy Unit GitHub repo}
\usage{
get_su_gh_file(repo, folder, file)
}
\arguments{
\item{repo}{string. The name of the repository in which to find the file}

\item{folder}{string. The folder where the file is located. Set to \code{""} to
use the root folder of the repo.}

\item{file}{string. The name of the file to read in}
}
\value{
The URL to the raw file contents, to be passed to a reader function
}
\description{
Read in a file from a Strategy Unit GitHub repo
}
\keyword{internal}
