% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpma_label_lookup.R
\name{get_tpma_label_lookup}
\alias{get_tpma_label_lookup}
\title{Prepare a lookup table for all current TPMAs}
\usage{
get_tpma_label_lookup()
}
\value{
A 4-column tibble, with columns \code{strategy}, \code{activity_type},
\code{change_factor} and \code{tpma_label}
}
\description{
Prepare a lookup table for all current TPMAs
}
