% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_type_pod_measure_options.R
\name{list_to_tbl}
\alias{list_to_tbl}
\title{Helper function to extract the required data fields from a list (from YAML)}
\usage{
list_to_tbl(lst)
}
\description{
Helper function to extract the required data fields from a list (from YAML)
}
\keyword{internal}
