% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpscenarioanalysis-package.R
\docType{package}
\name{nhpscenarioanalysis-package}
\alias{nhpscenarioanalysis}
\alias{nhpscenarioanalysis-package}
\title{nhpscenarioanalysis: Scenario result comparison app}
\description{
A Shiny app to compare model results across two different scenarios.
}
\author{
\strong{Maintainer}: Gabriel Hobro \email{gabriel.hobro@nhs.net}

Authors:
\itemize{
  \item Gabriel Hobro \email{gabriel.hobro@nhs.net}
  \item Ozayr Mohammed \email{ozayr.mohammed@nhs.net}
}

Other contributors:
\itemize{
  \item Zoë Turner \email{zoe.turner3@nhs.net} [contributor]
  \item Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID}) [contributor]
}

}
\keyword{internal}
