% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{read_pods_lookup}
\alias{read_pods_lookup}
\title{Read the pods lookup table from a YAML file in the NHP Outputs repo}
\usage{
read_pods_lookup(file = "golem-config.yml")
}
\description{
Read the pods lookup table from a YAML file in the NHP Outputs repo
}
